+++ MonsterVariety/ModEntry.cs
@@ -1,7 +1,6 @@
 using StardewModdingAPI;
 using StardewValley;
-using StardewValley.Delegates;
-using StardewValley.Extensions;
+using StardewValley.Internal;
 
 namespace MonsterVariety;
 
@@ -24,56 +23,9 @@ public class ModEntry : Mod
         ManageVariety.Apply(helper);
         VanillaCharacterMonster = helper.Data.ReadJsonFile<HashSet<string>>("assets/vanilla_character_monsters.json");
 
-        GameStateQuery.Register($"{ModId}_LUCKY_RANDOM", LUCKY_RANDOM);
-        GameStateQuery.Register($"{ModId}_SYNCED_LUCKY_RANDOM", SYNCED_LUCKY_RANDOM);
-    }
-
-    public static bool RandomImpl(Random random, string[] query, int skipArguments)
-    {
-        if (!ArgUtility.TryGetFloat(query, skipArguments, out float valueFlt, out string error, "float chance"))
-        {
-            Log(error);
-            return false;
-        }
-        double value = valueFlt;
-        bool addDailyLuck = false;
-        float addPlayerLuck = 0;
-        for (int i = skipArguments + 1; i < query.Length; i++)
-        {
-            if (query[i].EqualsIgnoreCase("@addDailyLuck"))
-            {
-                addDailyLuck = true;
-            }
-            if (query[i].EqualsIgnoreCase("@addPlayerLuck"))
-            {
-                ArgUtility.TryGetOptionalFloat(query, i + 1, out addPlayerLuck, out _, 1f, "float playerLuckMod");
-            }
-        }
-        if (addDailyLuck)
-        {
-            value += Game1.player.DailyLuck;
-        }
-        value += addPlayerLuck / 100 * Game1.player.LuckLevel;
-        return random.NextDouble() < (double)value;
-    }
-
-    private bool LUCKY_RANDOM(string[] query, GameStateQueryContext context)
-    {
-        return RandomImpl(context.Random, query, 1);
-    }
-
-    private bool SYNCED_LUCKY_RANDOM(string[] query, GameStateQueryContext context)
-    {
-        if (
-            !ArgUtility.TryGet(query, 1, out var value, out var error, allowBlank: true, "string interval")
-            || !ArgUtility.TryGet(query, 2, out var value2, out error, allowBlank: true, "string key")
-            || !Utility.TryCreateIntervalRandom(value, value2, out Random random, out error)
-        )
-        {
-            Log(error, LogLevel.Error);
-            return false;
-        }
-        return RandomImpl(random, query, 2);
+        GameStateQuery.Register($"{ModId}_LUCKY_RANDOM", GameDelegates.LUCKY_RANDOM);
+        GameStateQuery.Register($"{ModId}_SYNCED_LUCKY_RANDOM", GameDelegates.SYNCED_LUCKY_RANDOM);
+        ItemQueryResolver.Register($"{ModId}_MONSTER_DROPS", GameDelegates.MONSTER_DROPS);
     }
 
     /// <summary>SMAPI static monitor Log wrapper</summary>