+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2025-11-19
+
+### Added
+- New item query `mushymato.MonsterVariety_MONSTER_DROPS`
+- `AlwaysOverride` is now a model that individually controls custom monster class / custom texture checking. Previous bool values still accepted.
+
 ## [0.6.0] - 2025-09-28
 
 ### Added